Renamed directLineConnected to botConnected and added allImagesLoaded
