+++ __tests__/basic.js
@@ -2,10 +2,9 @@ import { By, Key } from 'selenium-webdriver';
 
 import { imageSnapshotOptions, timeouts } from './constants.json';
 
-import allImagesLoaded from './setup/conditions/allImagesLoaded.js';
-import directLineConnected from './setup/conditions/directLineConnected';
-import minNumActivitiesReached from './setup/conditions/minNumActivitiesReached';
-import webChatLoaded from './setup/conditions/webChatLoaded';
+import allImagesLoaded from './setup/conditions/allImagesLoaded';
+import botConnected from './setup/conditions/botConnected';
+import minNumActivitiesShown from './setup/conditions/minNumActivitiesShown';
 
 // selenium-webdriver API doc:
 // https://seleniumhq.github.io/selenium/docs/api/javascript/module/selenium-webdriver/index_exports_WebDriver.html
@@ -13,14 +12,13 @@ import webChatLoaded from './setup/conditions/webChatLoaded';
 test('setup', async () => {
   const { driver, pageObjects } = await setupWebDriver();
 
-  await driver.wait(webChatLoaded(), timeouts.navigation);
-  await driver.wait(directLineConnected(), timeouts.directLine);
+  await driver.wait(botConnected(), timeouts.directLine);
 
   const input = await driver.findElement(By.css('input[type="text"]'));
 
   await input.sendKeys('layout carousel', Key.RETURN);
-  await driver.wait(minNumActivitiesReached(3), timeouts.directLine);
-  await driver.wait(allImagesLoaded(), timeouts.fetch);
+  await driver.wait(minNumActivitiesShown(3), 2000);
+  await driver.wait(allImagesLoaded(), 2000);
 
   // Hide cursor before taking screenshot
   await pageObjects.hideCursor();
